Summary:
analysis: Log how long each high-level operation takes
analysis/util: Replace Multisets with Counters
analysis/util: New Cache class
analysis/type: Implement our own data type interface
analysis/type: Implement arrays
analysis/type: Implement bit fields
analysis/type: Implement enums
analysis/type: Implement function types
analysis/type: Implement pointers
analysis/util: New Lazy class
analysis/type: Add an interface for aggregate (struct/union) types
analysis/type: Implement structs
analysis/type: Implement unions
analysis/type: Implement typedefs
analysis/dataflow: Use BcpiType in PtrDomain
analysis: Use BcpiType in FieldReference
analysis: Use BcpiType in AccessPattern
analysis: Use Layout in the cost model/optimizer
analysis: Remove bcpi.Field
analysis: Pack fields in parallel
analysis: Compute costs for each offset in parallel
analysis: Remove remaining uses of Ghidra's type representation
analysis/type: Check structural equivalence before de-duplicating types

Test Plan:
./scripts/analyze.sh kernel.full address_info.csv
for csv in cpu2017/*.csv; ./scripts/analyze.sh ${csv%.csv} $csv

Results are the same before & after (except for minor differences in the
kernel analysis due to de-duplication fixes).

Reviewers: ali, oscar, pinosaur, ary, jiafengyu

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D842

+++ scripts/DebugAnalysis.java
@@ -57,7 +57,7 @@ public class DebugAnalysis extends BcpiAnalysis {
 		data.getRows()
 			.stream()
 			.filter(r -> refs.getFields(r.address).isEmpty())
-			.sorted(Comparator.<BcpiDataRow>comparingInt(r -> -r.getCount(BcpiConfig.CACHE_MISS_COUNTER)))
+			.sorted(Comparator.<BcpiDataRow>comparingLong(r -> -r.getCount(BcpiConfig.CACHE_MISS_COUNTER)))
 			.limit(10)
 			.forEach(r -> {
 				System.out.format("%s: %d samples, no struct fields\n",
